# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API para recomendar y buscar canciones en base a un dataset de características de canciones.",
    "version": "1.0.0",
    "title": "Spotify Recommender API",
    "contact": {
      "email": "contact@example.com"
    }
  },
  "host": "localhost:5000",
  "basePath": "/",
  "tags": [
    {
      "name": "track",
      "description": "Operaciones sobre canciones"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/recommend_track": {
      "get": {
        "tags": [
          "track"
        ],
        "summary": "Recomendar canciones",
        "description": "Recomienda canciones basadas en el nombre de una canción proporcionada y sus características numéricas.",
        "parameters": [
          {
            "name": "track_name",
            "in": "query",
            "description": "Nombre de la canción en la cual basar las recomendaciones",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Una lista de canciones recomendadas",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "artist": {
                    "type": "string"
                  },
                  "track_id": {
                    "type": "string"
                  },
                  "track_name": {
                    "type": "string"
                  },
                  "album": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error: track_name es requerido"
          },
          "404": {
            "description": "No se encontraron canciones"
          }
        }
      }
    },
    "/search_track": {
      "get": {
        "tags": [
          "track"
        ],
        "summary": "Buscar canciones",
        "description": "Busca canciones basadas en el nombre de una canción proporcionada.",
        "parameters": [
          {
            "name": "track_name",
            "in": "query",
            "description": "Nombre de la canción a buscar",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Una lista de canciones que coinciden con los criterios de búsqueda",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "artist": {
                    "type": "string"
                  },
                  "track_id": {
                    "type": "string"
                  },
                  "track_name": {
                    "type": "string"
                  },
                  "album": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error: track_name es requerido"
          },
          "404": {
            "description": "No se encontraron canciones"
          }
        }
      }
    }
  }
}